COPILOT-AGENT-1/ai-interview-agent-1/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
aiofiles
tempfile
base64
whisper
pyttsx3
vosk